RunFitControlWidget moved from FitActivityPanel down to FitSuiteWidget.
